requirements.txt
django==2.1.2
django-tenant-schemas==1.9.0
djangorestframework==3.9.0
